Construction of a common physics functions interface to replace the universal physics functions setup.
This commit introduces a new structure, the PhysicsFunctions structure in common_physics_functions.hpp.
This change is a reflection of conversations regarding the organization of the set of common physics functions
that we want to provide to all atmospheric processes.

Rather than adjust the current Functions structure defined in physics_functions.hpp we will construct a new
structure for all new functions.  This is mostly because the functions in physics_functions.hpp are already in use
throughout the code and any change would require adjustements throughout the code.  This can be done in future commits.

The first common function defined is the get_exner function.  There are also a series of property tests included in this commit.